Add files via upload
This has my full code, in which the glm seems to give high significance to all of the variables. I am not sure if this is a problem. But I was able to omit NAs in one iteration of this code and have a variety of significances. I am not sure what is causing this, but it still works as a model, and can make predictions.